requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.15.0
gunicorn==23.0.0
packaging==25.0
sqlparse==0.5.3
typing_extensions==4.13.1
django-allauth==65.7.0
requests==2.31.0
django-extensions>=3.2.0
werkzeug>=2.2.0
pyOpenSSL>=23.0.0